pyproject.toml
[tool.black]
line-length = 120
skip-string-normalization = '1'
exclude = '''
/(
  | env
  | proto
  |local_settings.py
)/
'''

[tool.poetry]
name = "alfa"
version = "0.3.0"
description = "alfa"
authors = ["Anton Ogorodnikov"]


[tool.poetry.dependencies]
"ruamel.yaml" = "^0.16.10"
aiohttp = "3.*" # for sentry
aiopg = "1.*"
coloredlogs = "14.*"
delorean = "1.*"
fastapi = "0.*"
h11 = "0.*"
httpx = "0.*"
jaeger-client = "4.*"
orjson = "3.*"
prometheus-client = "0.*"
python = "3.9.*"
python-multipart = "0.*" # for fastapi
sadisplay = "0.*"
sentry-sdk = "0.*"
sqlalchemy = "1.*"
uvicorn = "0.*"
statsd = "3.*"
pydantic = ">=1.5"
asyncpg = "^0.23.0"

[tool.poetry.dev-dependencies]
autoflake = "1.*"
black = "^19.10b0"
flake8 = "3.*"
isort = "4.*"
lxml = "4.*"
mypy = "0.790"
pip = "20.3"
poetry = "1.1.4"
pytest = "6.*"
pytest_httpx = "0.*"
pytest-asyncio = "0.*"
pytest-cov = "2.*"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
